scripts: Use qcom's merge_dtbs to compile dtbo.img and dtb
[For 5.10+ qcom kernels]
Required by the bootloader since the devicetrees in subdirectoires we have under the dts directory (qcom/$(folder/*)) are meant to be merged with devicetrees in qcom/* or the base dts directory.

Also use aosp's mkdtboimg.py to build dtbo.img

DTB is useful for anykernel zips, it'll be built whenever dtbo image is being built.

The path for dtb image is:-

$(objtree)/arch/$(SRCARCH)/boot/dtb

Signed-off-by: HELLBOY017 <abhaygill017@gmail.com>
Signed-off-by: BladeRunner-A2C <creator4983@aospa.co>